moved timetable feature for each stream platforms
