+++ common/src/main/java/com/freshdigitable/yttt/FetchStreamUseCase.kt
@@ -1,0 +1,5 @@
+package com.freshdigitable.yttt
+
+interface FetchStreamUseCase {
+    suspend operator fun invoke()
+}
