+++ common/src/main/java/com/freshdigitable/yttt/TimetablePage.kt
@@ -1,0 +1,10 @@
+package com.freshdigitable.yttt
+
+enum class TimetablePage(val type: Type = Type.SIMPLE) {
+    OnAir,
+    Upcoming(type = Type.GROUPED),
+    FreeChat,
+    ;
+
+    enum class Type { SIMPLE, GROUPED, }
+}
