+++ common/src/main/java/com/freshdigitable/yttt/di/TimetableTabModuleAnnotations.kt
@@ -1,0 +1,13 @@
+package com.freshdigitable.yttt.di
+
+import com.freshdigitable.yttt.TimetablePage
+import dagger.MapKey
+import javax.inject.Qualifier
+
+@Suppress("unused")
+@Qualifier
+annotation class TimetableTabQualifier(val page: TimetablePage)
+
+@Suppress("unused")
+@MapKey
+annotation class TimetableTabKey(val page: TimetablePage)
