+++ feature-timetable-youtube/src/main/java/com/freshdigitable/yttt/feature/timetable/youtube/di/YouTubeTimetableFeatureModule.kt
@@ -1,0 +1,48 @@
+package com.freshdigitable.yttt.feature.timetable.youtube.di
+
+import com.freshdigitable.yttt.FetchStreamUseCase
+import com.freshdigitable.yttt.FetchTimetableItemSourceUseCase
+import com.freshdigitable.yttt.TimetableContextMenuSelector
+import com.freshdigitable.yttt.TimetablePage
+import com.freshdigitable.yttt.data.model.YouTubeVideo
+import com.freshdigitable.yttt.di.IdBaseClassKey
+import com.freshdigitable.yttt.di.TimetableTabQualifier
+import com.freshdigitable.yttt.feature.timetable.youtube.FetchYouTubeFreeChatItemSourceUseCase
+import com.freshdigitable.yttt.feature.timetable.youtube.FetchYouTubeOnAirItemSourceUseCase
+import com.freshdigitable.yttt.feature.timetable.youtube.FetchYouTubeStreamUseCase
+import com.freshdigitable.yttt.feature.timetable.youtube.FetchYouTubeUpcomingItemSourceUseCase
+import com.freshdigitable.yttt.feature.timetable.youtube.TimetableContextMenuDelegateForYouTube
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ViewModelComponent
+import dagger.multibindings.IntoMap
+import dagger.multibindings.IntoSet
+
+@InstallIn(ViewModelComponent::class)
+@Module
+internal interface YouTubeTimetableFeatureModule {
+    @Binds
+    @IntoSet
+    fun bindYouTubeStreamUseCase(useCase: FetchYouTubeStreamUseCase): FetchStreamUseCase
+
+    @Binds
+    @IntoMap
+    @IdBaseClassKey(YouTubeVideo.Id::class)
+    fun bindTimetableContextMenuDelegateForYouTube(delegate: TimetableContextMenuDelegateForYouTube): TimetableContextMenuSelector
+
+    @Binds
+    @IntoSet
+    @TimetableTabQualifier(TimetablePage.OnAir)
+    fun bindFetchYouTubeOnAirItemSourceUseCase(useCase: FetchYouTubeOnAirItemSourceUseCase): FetchTimetableItemSourceUseCase
+
+    @Binds
+    @IntoSet
+    @TimetableTabQualifier(TimetablePage.Upcoming)
+    fun bindFetchYouTubeUpcomingItemSourceUseCase(useCase: FetchYouTubeUpcomingItemSourceUseCase): FetchTimetableItemSourceUseCase
+
+    @Binds
+    @IntoSet
+    @TimetableTabQualifier(TimetablePage.FreeChat)
+    fun bindFetchYouTubeFreeChatItemSourceUseCase(useCase: FetchYouTubeFreeChatItemSourceUseCase): FetchTimetableItemSourceUseCase
+}
