+++ lib/src/main/java/com/freshdigitable/yttt/TimetablePageDelegate.kt
@@ -2,20 +2,14 @@ package com.freshdigitable.yttt
 
 import com.freshdigitable.yttt.compose.TabData
 import com.freshdigitable.yttt.data.SettingRepository
-import com.freshdigitable.yttt.data.TwitchLiveRepository
-import com.freshdigitable.yttt.data.YouTubeRepository
 import com.freshdigitable.yttt.data.model.LiveVideo
-import com.freshdigitable.yttt.data.model.TwitchUserDetail
 import com.freshdigitable.yttt.data.model.dateWeekdayFormatter
-import com.freshdigitable.yttt.data.model.toLiveVideo
 import com.freshdigitable.yttt.data.model.toLocalDateTime
-import com.freshdigitable.yttt.data.model.toTwitchVideoList
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.map
 import java.time.Duration
-import java.time.Instant
 import java.time.temporal.ChronoUnit
 import javax.inject.Inject
 
@@ -23,60 +17,6 @@ interface TimetablePageDelegate {
     fun getSimpleItemList(page: TimetablePage): Flow<List<LiveVideo>>
     fun getGroupedItemList(page: TimetablePage): Flow<Map<String, List<LiveVideo>>>
     val tabs: Flow<List<TabData>>
-}
-
-interface FetchTimetableItemSourceUseCase {
-    operator fun invoke(): Flow<List<LiveVideo>>
-}
-
-class FetchTwitchOnAirItemSourceUseCase @Inject constructor(
-    private val repository: TwitchLiveRepository,
-) : FetchTimetableItemSourceUseCase {
-    override operator fun invoke(): Flow<List<LiveVideo>> = repository.onAir.map {
-        it.map { s ->
-            val user = s.user as? TwitchUserDetail
-                ?: repository.findUsersById(listOf(s.user.id)).first()
-            s.toLiveVideo(user)
-        }
-    }
-}
-
-class FetchYouTubeOnAirItemSourceUseCase @Inject constructor(
-    private val repository: YouTubeRepository,
-) : FetchTimetableItemSourceUseCase {
-    override fun invoke(): Flow<List<LiveVideo>> = repository.videos.map { v ->
-        v.filter { it.isNowOnAir() }.map { it.toLiveVideo() }
-    }
-}
-
-class FetchYouTubeUpcomingItemSourceUseCase @Inject constructor(
-    private val repository: YouTubeRepository,
-) : FetchTimetableItemSourceUseCase {
-    override fun invoke(): Flow<List<LiveVideo>> = repository.videos.map { v ->
-        v.filter { it.isUpcoming() && it.isFreeChat != true }.map { it.toLiveVideo() }
-    }
-}
-
-class FetchTwitchUpcomingItemSourceUseCase @Inject constructor(
-    private val repository: TwitchLiveRepository,
-) : FetchTimetableItemSourceUseCase {
-    override fun invoke(): Flow<List<LiveVideo>> = repository.upcoming.map { u ->
-        val week = Instant.now().plus(Duration.ofDays(7L))
-        u.map { s -> s.toTwitchVideoList() }.flatten()
-            .filter { it.schedule.startTime.isBefore(week) }
-            .map { s ->
-                val user = s.user as? TwitchUserDetail
-                    ?: repository.findUsersById(listOf(s.user.id)).first()
-                s.toLiveVideo(user)
-            }
-    }
-}
-
-class FetchYouTubeFreeChatItemSourceUseCase @Inject constructor(
-    private val repository: YouTubeRepository,
-) : FetchTimetableItemSourceUseCase {
-    override fun invoke(): Flow<List<LiveVideo>> =
-        repository.videos.map { v -> v.filter { it.isFreeChat == true }.map { it.toLiveVideo() } }
 }
 
 class TimetablePageDelegateImpl @Inject constructor(
