+++ lib/src/main/java/com/freshdigitable/yttt/di/TimetableTabModule.kt
@@ -1,30 +1,14 @@
 package com.freshdigitable.yttt.di
 
 import com.freshdigitable.yttt.FetchTimetableItemSourceUseCase
-import com.freshdigitable.yttt.FetchTwitchOnAirItemSourceUseCase
-import com.freshdigitable.yttt.FetchTwitchUpcomingItemSourceUseCase
-import com.freshdigitable.yttt.FetchYouTubeFreeChatItemSourceUseCase
-import com.freshdigitable.yttt.FetchYouTubeOnAirItemSourceUseCase
-import com.freshdigitable.yttt.FetchYouTubeUpcomingItemSourceUseCase
 import com.freshdigitable.yttt.TimetablePage
 import com.freshdigitable.yttt.TimetablePageDelegate
 import com.freshdigitable.yttt.TimetablePageDelegateImpl
 import dagger.Binds
-import dagger.MapKey
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ViewModelComponent
 import dagger.multibindings.IntoMap
-import dagger.multibindings.IntoSet
-import javax.inject.Qualifier
-
-@Suppress("unused")
-@Qualifier
-annotation class TimetableTabQualifier(val page: TimetablePage)
-
-@Suppress("unused")
-@MapKey
-annotation class TimetableTabKey(val page: TimetablePage)
 
 @Suppress("unused")
 @InstallIn(ViewModelComponent::class)
@@ -53,29 +37,4 @@ interface TimetableTabModules {
 
     @Binds
     fun bindDelegate(facade: TimetablePageDelegateImpl): TimetablePageDelegate
-
-    @Binds
-    @IntoSet
-    @TimetableTabQualifier(TimetablePage.OnAir)
-    fun bindFetchTwitchOnAirItemSourceUseCase(useCase: FetchTwitchOnAirItemSourceUseCase): FetchTimetableItemSourceUseCase
-
-    @Binds
-    @IntoSet
-    @TimetableTabQualifier(TimetablePage.OnAir)
-    fun bindFetchYouTubeOnAirItemSourceUseCase(useCase: FetchYouTubeOnAirItemSourceUseCase): FetchTimetableItemSourceUseCase
-
-    @Binds
-    @IntoSet
-    @TimetableTabQualifier(TimetablePage.Upcoming)
-    fun bindFetchTwitchUpcomingItemSourceUseCase(useCase: FetchTwitchUpcomingItemSourceUseCase): FetchTimetableItemSourceUseCase
-
-    @Binds
-    @IntoSet
-    @TimetableTabQualifier(TimetablePage.Upcoming)
-    fun bindFetchYouTubeUpcomingItemSourceUseCase(useCase: FetchYouTubeUpcomingItemSourceUseCase): FetchTimetableItemSourceUseCase
-
-    @Binds
-    @IntoSet
-    @TimetableTabQualifier(TimetablePage.FreeChat)
-    fun bindFetchYouTubeFreeChatItemSourceUseCase(useCase: FetchYouTubeFreeChatItemSourceUseCase): FetchTimetableItemSourceUseCase
 }
