+++ settings.gradle
@@ -16,6 +16,9 @@ rootProject.name = "yttt"
 include ':app'
 include ':lib'
 include ':common'
+include ':feature-timetable-youtube'
+include ':feature-timetable-twitch'
 include ':repository-youtube'
 include ':repository-twitch'
+include ':repository-appuser'
 include ':data-local-room'
